feat: SEO improvements - Added SEO metadata for events, projects, and tweets pages - Updated robots.txt and sitemap.xml for better indexing - Improved overall SEO structure and organization
