+++ src/app/projects/page.tsx
@@ -1,35 +1,9 @@
 import { Metadata } from 'next';
 import { getPublishedRepos } from '@/actions/projects';
+import { projectsMetadata } from '@/config/seo/projects-metadata';
 
-export const metadata: Metadata = {
-  title: 'Projects | GDSC NIT Rourkela',
-  description:
-    'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
-  keywords: [
-    'GDSC',
-    'NIT Rourkela',
-    'projects',
-    'open source',
-    'student developers',
-    'GitHub',
-  ],
-  openGraph: {
-    title: 'Projects | GDSC NIT Rourkela',
-    description:
-      'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
-    images: ['/opengraph-image.png'],
-    type: 'website',
-  },
-  twitter: {
-    card: 'summary_large_image',
-    title: 'Projects | GDSC NIT Rourkela',
-    description:
-      'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
-    images: ['/opengraph-image.png'],
-  },
-};
+export const metadata: Metadata = projectsMetadata;
 
 export default function ProjectsPage() {
-  // Projects page implementation
-  return <div>{/* Projects listing */}</div>;
+  return <div></div>;
 }
