+++ src/config/seo/projects-metadata.ts
@@ -1,0 +1,29 @@
+import { Metadata } from 'next';
+
+export const projectsMetadata: Metadata = {
+  title: 'Projects | GDSC NIT Rourkela',
+  description:
+    'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
+  keywords: [
+    'GDSC',
+    'NIT Rourkela',
+    'projects',
+    'open source',
+    'student developers',
+    'GitHub',
+  ],
+  openGraph: {
+    title: 'Projects | GDSC NIT Rourkela',
+    description:
+      'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
+    images: ['/opengraph-image.png'],
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Projects | GDSC NIT Rourkela',
+    description:
+      'Explore innovative projects developed by Google Developer Student Club at NIT Rourkela',
+    images: ['/opengraph-image.png'],
+  },
+};
